Update README.md
Added an About tab and more content to the contributions part of the readme. This is so when users and contributors access the repository they have a better idea about the contents of the project. 
The About tab highlights the features of ToolZar
The contributions tab highlights how a contributor can get started and which folders are best to look into when adding contributions to the code.